Directly using x11, takp_launcher.conf, and also using inotify
Lots of fixes. We don't shell out to find and raise windows anymore, preferring
directly calling X11 instead.
Parsing the config means I don't have to specify account names on the commandline.
And using inotify to re-parse the config when it changes means I don't have to
relaunch the program to change the accounts anymore.